mini-library-management/backend/requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.19.2
mysql-connector-python==8.0.27
pytest==6.2.5
requests==2.26.0
Werkzeug==2.0.3
SQLAlchemy<1.4